fix start of lsp service
