+++ server/package.json
@@ -22,7 +22,6 @@
   "dependencies": {
     "@prisma/get-platform": "2.0.0-alpha.1306",
     "pkg-dir": "4.2.0",
-    "prismafile": "1.0.3",
     "vscode-languageserver": "6.1.1",
     "vscode-languageserver-textdocument": "1.0.1",
     "vscode-uri": "2.1.1",
